docs: landing snippets as headers
