+++ test-files/golden-tests/snippets/distance.adoc
@@ -8,7 +8,7 @@ Return the distance between two points
 
 === Synopsis
 
-Declared in `&lt;distance&period;cpp&gt;`
+Declared in `&lt;distance&period;hpp&gt;`
 
 [source,cpp,subs="verbatim,replacements,macros,-callouts"]
 ----
