+++ test-files/golden-tests/snippets/distance.cpp
@@ -1,13 +1,1 @@
-/** Return the distance between two points
-
-    This function returns the distance between two points
-    according to the Euclidean distance formula.
-
-    @param x0 The x-coordinate of the first point
-    @param y0 The y-coordinate of the first point
-    @param x1 The x-coordinate of the second point
-    @param y1 The y-coordinate of the second point
-    @return The distance between the two points
-*/
-double
-distance(double x0, double y0, double x1, double y1);
+#include "distance.hpp"
