+++ test-files/golden-tests/snippets/function_object.cpp
@@ -1,13 +1,1 @@
-struct abs_fn
-{
-    /** Compute the absolute value.
-
-        @param x The input value.
-        @return The absolute value of x.
-    */
-    double
-    operator()(double x) const noexcept;
-};
-
-/** Return the absolute value of a number. */
-constexpr abs_fn abs = {};
+#include "function_object.hpp"
