+++ test-files/golden-tests/snippets/function_object.hpp
@@ -1,0 +1,13 @@
+struct abs_fn
+{
+    /** Compute the absolute value.
+
+        @param x The input value.
+        @return The absolute value of x.
+    */
+    double
+    operator()(double x) const noexcept;
+};
+
+/** Return the absolute value of a number. */
+constexpr abs_fn abs = {};
