+++ test-files/golden-tests/snippets/is_prime.adoc
@@ -8,7 +8,7 @@ Return true if a number is prime&period;
 
 === Synopsis
 
-Declared in `&lt;is&lowbar;prime&period;cpp&gt;`
+Declared in `&lt;is&lowbar;prime&period;hpp&gt;`
 
 [source,cpp,subs="verbatim,replacements,macros,-callouts"]
 ----
