+++ test-files/golden-tests/snippets/is_prime.cpp
@@ -1,11 +1,1 @@
-/** Return true if a number is prime.
-
-    @par Complexity
-
-    Linear in n.
-
-    @return Whether `n` is prime.
-    @param n The number to test
-*/
-bool
-is_prime(unsigned long long n) noexcept;
+#include "is_prime.hpp"
