+++ test-files/golden-tests/snippets/terminate.cpp
@@ -1,9 +1,1 @@
-/** Exit the program.
-
-    The program will end immediately.
-
-    @note This function does not return.
-*/
-[[noreturn]]
-void
-terminate() noexcept;
+#include "terminate.hpp"
