+++ test-files/golden-tests/snippets/terminate.hpp
@@ -1,0 +1,9 @@
+/** Exit the program.
+
+    The program will end immediately.
+
+    @note This function does not return.
+*/
+[[noreturn]]
+void
+terminate() noexcept;
